LocationManager: expose more information from the backend to the apps
note: GnssStatus was accidentally bundled into ab5b600bf151fa9320052a6fa8cbdf9c238353ca

+++ src/api-impl/android/hardware/SensorManager.java
@@ -23,7 +23,7 @@ public class SensorManager {
 				new LocationManager().requestLocationUpdates(null, 0, 0, new LocationListener() {
 					@Override
 					public void onLocationChanged(Location location) {
-						listener.onSensorChanged(new SensorEvent(new float[]{(float)location.getBearing()}, sensor));
+						listener.onSensorChanged(new SensorEvent(new float[]{location.getBearing()}, sensor));
 					}
 				});
 				return true;
